Link to front end for payment page and fixing total amount logic
If the cart is empty and user brute force to payment page, it will have an alert message and brute force them back to shopping page. The payment total even if inspect element change the amount to smaller, the database will still take in the correct total amount to be paid.